requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.4
requests==2.25.1
gunicorn==20.1.0
python-dotenv==1.0.0